Refactored to seperate routes and queries
